# backend/openapi.json
{
  "components": {
    "schemas": {
...
      "ErrorBody": {
        "description": "Typed error payload: ``error_type`` is the stable taxonomy string the\nUI maps onto actions (runbook link, retry button, hard stop).",
        "properties": {
          "detail": {
            "title": "Detail",
            "type": "string"
          },
          "error_type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "error_type",
          "detail"
        ],
        "title": "ErrorBody",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "InviteCreate": {
        "description": "POST /api/admin/invites body. Lightweight email shape check (a friends-\ninvite flow, not a validation fortress) — normalization happens server-side.",
        "properties": {
          "email": {
            "maxLength": 320,
            "minLength": 3,
            "title": "Email",
            "type": "string"
          }
        },
        "required": [
          "email"
        ],
        "title": "InviteCreate",
        "type": "object"
      },
      "InviteList": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/InviteOut"
            },
            "title": "Items",
            "type": "array"
          }
        },
        "required": [
          "items"
        ],
        "title": "InviteList",
        "type": "object"
      },
      "InviteOut": {
        "description": "An invite as the admin sees it — NEVER the token_hash. ``dev_activation_\nlink`` is present ONLY when chefclaw_email='fake' (the real link is emailed).",
        "properties": {
          "accepted_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Accepted At"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "dev_activation_link": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dev Activation Link"
          },
          "email": {
            "title": "Email",
            "type": "string"
          },
          "expires_at": {
            "format": "date-time",
            "title": "Expires At",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "id",
          "email",
          "status",
          "expires_at",
          "created_at"
        ],
        "title": "InviteOut",
        "type": "object"
      },
      "InvitePublicOut": {
        "description": "GET /api/invites/{token} — the public invite-accept shape (M13). ``status``\nis 'pending' | 'invalid'; ``email`` is revealed ONLY for a live pending\ninvite (a missing/expired/revoked token is a uniform 'invalid', no email).",
        "properties": {
          "email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email"
          },
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "status"
        ],
        "title": "InvitePublicOut",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "chefclaw",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/admin/invites": {
      "get": {
        "description": "List invites (newest first), optionally filtered by status. NEVER returns\na token_hash.",
        "operationId": "list_invites_api_admin_invites_get",
        "parameters": [
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InviteList"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Invites",
        "tags": [
          "invites"
        ]
      },
      "post": {
        "description": "Issue (or rotate + resend) a pending invite. Already an active member ⇒\n409; a new invite ⇒ 201; a rotate/resend ⇒ 200. The activation link is\nemailed; it rides in the response ONLY when chefclaw_email='fake'.",
        "operationId": "create_invite_api_admin_invites_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InviteCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InviteOut"
                }
              }
            },
            "description": "OK"
          },
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InviteOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "502": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Bad Gateway"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Create Invite",
        "tags": [
          "invites"
        ]
      }
    },
    "/api/admin/invites/{invite_id}/revoke": {
      "post": {
        "description": "Revoke a pending invite (idempotent). Already-accepted ⇒ 409; missing ⇒\n404; otherwise 200.",
        "operationId": "revoke_invite_api_admin_invites__invite_id__revoke_post",
        "parameters": [
          {
            "in": "path",
            "name": "invite_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Invite Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Revoke Invite",
        "tags": [
          "invites"
        ]
      }
    },
...
    "/api/invites/{token}": {
      "get": {
        "description": "PUBLIC invite-accept lookup (M13): a live pending invite reveals its\nemail; a missing/expired/revoked/accepted token is a uniform 'invalid' with\nno email (no enumeration oracle, no address leak).",
        "operationId": "public_invite_api_invites__token__get",
        "parameters": [
          {
            "in": "path",
            "name": "token",
            "required": true,
            "schema": {
              "title": "Token",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvitePublicOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Public Invite",
        "tags": [
          "invites"
        ]
      }
    },
...
  }
}